cast to int for float labels

+++ creator_lora/utils/sampler_weights.py
@@ -1,7 +1,9 @@
 import torch
 
 def compute_sampler_weights(labels):
-    label_counts = torch.bincount(torch.tensor(labels))
+    # manually cast to int for float labels for approximate weights
+    # I hope it would be good enough
+    label_counts = torch.bincount(torch.tensor(labels).int())
     label_weights = 1 / label_counts.float()
     weights = label_weights[labels]
     weights /= weights.sum()
